requirements.txt
pytest==8.4.1
requests==2.32.3
allure-pytest==2.14.3
PyYAML==6.0.1
